New diagonal version and set Box's parent

+++ Assets/Scripts/NewObjectPanel.cs
@@ -43,7 +43,7 @@ public class NewObjectPanel : MonoBehaviour
         HidePanelControl();
     }
 
-    public void ChangeSize(string axis, int value)
+    public void ChangeSize(string axis, float value)
     {
         switch (axis)
         {
